requirements.txt
python-dotenv
agent-framework[azure-ai]
azure-identity
pydantic
pandas
fastapi
uvicorn
httpx
pyyaml
